mod/#162 산책 시작 API issuedAt 타입을 Long으로 변경 및 deviceInfo를 필수 값으로 수정

+++ app/src/main/java/com/paw/key/data/dto/response/walk/WalkStartResponseDto.kt
@@ -9,7 +9,7 @@ data class WalkStartResponseDto(
     @SerialName("routeId")
     val routeId: String,
     @SerialName("issuedAt")
-    val issuedAt: Int,
+    val issuedAt: Long,
 ) {
     fun toEntity() = WalkStartEntity(
         routeId = routeId,
